filter constructors
filter constructors now works as stand alone
