+++ src/components/map.jsx
@@ -1,96 +1,19 @@
 import React, { useState, useEffect } from "react";
 import { Chart } from "react-google-charts";
 import { Box } from "@mui/material";
-import axios from "axios";
-
-const url = "http://ergast.com/api/f1/constructors.json?limit=300";
-
-const useFetchData = (url) => {
-  const [data, setData] = useState(null);
-  
-  useEffect(() => {
-    const fetchData = async () => {
-      try {
-        const response = await axios.get(url);
-        setData(response.data);
-      } catch (error) {
-        console.error(error);
-      }
-    };
-    
-    fetchData();
-  }, []);
-  
-  return data;
-};
+import filterConstructors from '../hooks/filterConstructors.js';
 
 export default function Map() {
-  const data = useFetchData(url);
-  const [nationalities, setNationalities] = useState([]);
-  const [uniqueNationalities, setUniqueNationalities] = useState([]);
   const [nationalityCount, setNationalityCount] = useState([]);
-  const countryNames = 
-  [
-    'United States',
-    'Germany',
-    'France',
-    'Switzerland',
-    'Italy',
-    'United Kingdom',
-    'New Zealand',
-    'Netherlands',
-    'Malaysia',
-    'Germany',
-    'Belgium',
-    'Brazil',
-    'India',
-    'Japan',
-    'Spain',
-    'Ireland',
-    'South Africa',
-    'Russia',
-    'Australia',
-    'Rhodesia',
-    'Mexico',
-    'Austria',
-    'Canada',
-    'China'
-  ];
 
   useEffect(() => {
-    if (data) {
-      const constructorData = data.MRData.ConstructorTable.Constructors;
-      if (constructorData) {
-        const nationalitiesArray = constructorData.map(
-          (constructor) => constructor.nationality
-        );
-        setNationalities(nationalitiesArray);
-      }
-    }
-  }, [data]);
+    const fetchData = async () => {
+      const result = await filterConstructors();
+      setNationalityCount(result);
+    };
 
-  useEffect(() => {
-    if (nationalities.length > 0) {
-      const uniqueNationalitiesSet = new Set(nationalities);
-      setUniqueNationalities(Array.from(uniqueNationalitiesSet));
-    }
-  }, [nationalities.length]);
-
-  useEffect(() => {
-    if (uniqueNationalities.length > 0) {
-      const header = ["Country", "Number of constructors from each country"];
-      let countArray = uniqueNationalities.map((uniqueNationality, index) => ([
-        countryNames[index],
-        nationalities.filter((n) => n === uniqueNationality).length,
-      ]));
-      countArray.unshift(header);
-      countArray[1][1]++;
-      countArray = countArray.filter((_, index) => index !== 10);
-      setNationalityCount(countArray);
-    }
-  }, [uniqueNationalities.length]);
-
-  console.log(nationalityCount);
+    fetchData();
+  }, []);
 
   return (
     <Box sx={{
